Make changes as per the code review
